recommendation_service/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
PyJWT>=2.8
redis>=5.0
requests>=2.31
celery>=5.3
scikit-learn>=1.4
pandas>=2.2
numpy>=1.26

gunicorn>=21.2.0

whitenoise>=6.6.0

whitenoise>=6.6.0
